+++ src/xSTIR/pSTIR/tests/test_ObjectiveFunction.py
@@ -75,3 +75,23 @@ class TestSTIRObjectiveFunction(unittest.TestCase):
         print('relative difference: %f' % q)
         assert q <= .002
 
+    def test_gradient_inplace(self):
+        '''Checks that gradient computed in-place returns the same as without in-place.
+
+        Note: I had to increase the absolute tolerance of the test from the default 1e-8 to 1e-5, see 'test_gradient_out'
+        '''
+        x = self.image
+        g1 = self.obj_fun.gradient(x)
+        self.obj_fun.gradient(x, out=x)
+        numpy.testing.assert_allclose(x.asarray(), g1.asarray(), atol=1e-5)
+
+    def test_gradient_out(self):
+        '''Checks that gradient with 'out' parameter returns the same as without the 'out' parameter.
+
+        Note: I had to increase the absolute tolerance of the test from the default 1e-8 to 1e-5.
+        '''
+        x = self.image
+        g1 = self.obj_fun.gradient(x)
+        g2 = g1 * 0
+        self.obj_fun.gradient(x, out=g2)
+        numpy.testing.assert_allclose(g2.asarray(), g1.asarray(), atol=1e-5)